# src/main/resources/META-INF/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Provoly Hypervisor API
  description: CRUD Operation on repositories data
  version: 1.0.0
servers:
  - url: http://provoly-hypervisor.com
paths:
...
  /procedures:
    post:
      tags:
        - Procedure Controller
      summary: Update events procedure
      operationId: com.provoly.event.ProcedureController_updateProcedure
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcedureWriteDto'
            examples:
              default:
                summary: Update Report event from procedure
                value: |
                  {
                    "id": "f7b37e0e-a3d1-4159-a2c1-f6b7fd1f7a82",
                    "name": "procedure_1",
                    "events": [
                      {
                        "id": "ebbfbbd1-b8de-467c-8a9b-88443a49f807",
                        "name": "Accident voiture UPDATE",
                        "address": "10 Rue du Camp d'Attila, 51000 Châlons-en-Champagne",
                        "description": "description",
                        "criticality": "HIGH",
                        "type": "REPORT",
                        "equipmentId": "4ead3c85-f120-4144-8ca2-29ac9882ff5f",
                        "externalSourceRef": "external_source2",
                        "category": "REPORT"
                      }
                    ]
                  }
        

      responses:
        "201":
          description: Procedure Created
        "204":
          description: Updated
        "400":
          description: Bad Request
        "401":
          description: Not Authorized
      security:
        - oidc:
            - authenticated
  /procedures/id/{id}:
    get:
      tags:
        - Procedure Controller
      summary: Retrieve procedure by id
      operationId: com.provoly.event.ProcedureController_getProcedureDetails
      parameters:
        - name: id
          in: path
          schema:
            $ref: '#/components/schemas/UUID'
          example: 01ffde9d-30d2-4273-b61f-c6addd8747c8
          required: true
      responses:
        "401":
          description: Not Authorized
        "403":
          description: Forbidden if page not specified
        "404":
          description: Procedure Not found
        "200":
          description: Returns procedure Details by id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcedureReadDto'
              examples:
                default:
                  value: |
                    {
                      "id": "5c925baf-7164-4bfe-af37-f0860aacc570",
                      "name": "procedure maintenance",
                      "creationDate": "2024-02-23T12:56:00.802501Z",
                      "actions": [
                        {
                          "id": "6b5ca865-87e5-467b-a996-4c0629e8f94c",
                          "name": "intervention procedure_2",
                          "type": "INTERVENTION_REQUEST",
                          "status": "IN_PROGRESS",
                          "lastModificationDate": "2024-02-23T12:56:00.802501Z"
                        },
                        {
                          "id": "a08e2110-3a9c-4f77-bf65-31150201bbfa",
                          "name": "intervention procedure_2 bis",
                          "type": "INTERVENTION_REQUEST",
                          "status": "DONE",
                          "lastModificationDate": "2024-02-23T12:56:00.802501Z"
                        }
                      ],
                      "events": [
                        {
                          "id": "01ffde9d-30d2-4273-b61f-c6addd8747c8",
                          "name": "Defaillance lampe",
                          "address": null,
                          "description": null,
                          "criticality": "LOW",
                          "status": "DONE",
                          "type": "ALERT",
                          "lastModificationDate": "2024-02-23T12:56:00.802501Z",
                          "creationDate": "2024-02-23T12:56:00.802501Z",
                          "category": "ALERT_MALFUNCTION",
                          "equipment": {
                            "id": "12a64ae7-70ca-4d12-bcd9-9f1c75f23915",
                            "name": "C-1034",
                            "type": "Ouvrage",
                            "entity": "Châlons Agglo"
                          },
                          "externalSourceRef": "citylinx_1"
                        },
                        {
                          "id": "eb8e2f6b-f33e-408c-bb6c-e042f204d687",
                          "name": "Maintenance ouvrage2",
                          "address": null,
                          "description": null,
                          "criticality": "MEDIUM",
                          "status": "NEW",
                          "type": "REPORT",
                          "lastModificationDate": "2024-02-23T12:56:00.802501Z",
                          "creationDate": "2024-02-23T12:56:00.802501Z",
                          "category": "REPORT",
                          "equipment": {
                            "id": "12a64ae7-70ca-4d12-bcd9-9f1c75f23916",
                            "name": "C-761",
                            "type": "Ouvrage",
                            "entity": "Châlons Agglo"
                          },
                          "externalSourceRef": "external_source3"
                        }
                      ],
                      "progress": 50.0
                    }
      security:
        - oidc:
            - authenticated
  /procedures/id/{id}/close:
    put:
      tags:
        - Procedure Controller
      summary: Close all procedure events by id
      description: Set status to DONE. Close date is not updated if event is already closed
      operationId: com.provoly.event.EventController_closeEvent
      parameters:
        - name: id
          in: path
          schema:
            $ref: '#/components/schemas/UUID'
          example: 01ffde9d-30d2-4273-b61f-c6addd8747c8
          required: true
      responses:
        "204":
          description: Updated
        "401":
          description: Not Authorized
        "404":
          description: Event Not found
      security:
        - oidc:
            - authenticated
...
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      description: Authentication
      openIdConnectUrl: https://auth/realms/provoly/.well-known/openid-configuration
  schemas:
    UUID:
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      type: string
      example: "ebbfbbd1-b8de-467c-8a9b-88443a49f807"
      required: true
    DateTime:
      format: date-time
      type: string
      example: "2024-02-20T08:33:14.312632Z"
    Criticality:
      type: string
      required: true
      enum:
        - LOW
        - MEDIUM
        - HIGH
    Status:
      type: string
      required: true
      enum:
        - NEW
        - IN_PROGRESS
        - DONE
    Type:
      type: string
      required: true
      enum:
        - ALERT
        - OPERATOR
        - REPORT
    OperatorCategory:
      type: string
      required: true
      enum:
        - MANIFESTATION,
        - OPERATOR_EVENT
    AlertCategory:
      type: string
      required: true
      enum:
        - ALERT_LIMIT,
        - ALERT_MALFUNCTION
    ReportCategory:
      enum:
        - REPORT
...
    Family:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
        code:
          type: string
    EquipmentEntity:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
    EquipmentReadDto:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
        family:
          $ref: '#/components/schemas/Family'
        entity:
          $ref: '#/components/schemas/EquipmentEntity'
    ActionReadDto:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
        type:
          type: string
        status:
          $ref: '#/components/schemas/Status'
        lastModificationDate:
          $ref: '#/components/schemas/DateTime'
    OperatorEventWriteDto:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
          required: true
        address:
          type: string
          required: true
        description:
          type: string
          required: true
        criticality:
          $ref: '#/components/schemas/Criticality'
        category:
          $ref: '#/components/schemas/OperatorCategory'
        equipmentId:
          $ref: '#/components/schemas/UUID'
        endDate:
          $ref: '#/components/schemas/DateTime'
        startDate:
          $ref: '#/components/schemas/DateTime'
        type:
          type: string
          enum:
            - OPERATOR
          required: true
    AlertEventWriteDto:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
          required: true
        address:
          type: string
          required: true
        description:
          type: string
          required: true
        criticality:
          $ref: '#/components/schemas/Criticality'
        category:
          $ref: '#/components/schemas/AlertCategory'
        equipmentId:
          $ref: '#/components/schemas/UUID'
        externalSourceRef:
          type: string
          required: true
        type:
          type: string
          enum:
            - ALERT
          required: true
    ReportEventWriteDto:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
          required: true
        address:
          type: string
          required: true
        description:
          type: string
          required: true
        criticality:
          $ref: '#/components/schemas/Criticality'
        category:
          $ref: '#/components/schemas/ReportCategory'
        equipmentId:
          $ref: '#/components/schemas/UUID'
        externalSourceRef:
          type: string
          required: true
        type:
          type: string
          enum:
            - REPORT
          required: true
    AlertEventReadDto:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
        address:
          type: string
        description:
          type: string
        type:
          $ref: '#/components/schemas/Type'
        criticality:
          $ref: '#/components/schemas/Criticality'
        equipment:
          $ref: '#/components/schemas/EquipmentReadDto'
        status:
          $ref: '#/components/schemas/Status'
        category:
          $ref: '#/components/schemas/AlertCategory'
        lastModificationDate:
          $ref: '#/components/schemas/DateTime'
        creationDate:
          $ref: '#/components/schemas/DateTime'
        closeDate:
          $ref: '#/components/schemas/DateTime'
        externalSourceRef:
          type: string
        procedureId:
          $ref: '#/components/schemas/UUID'
        linkedEvents:
          type: number
          description: Number of events that have the same procedure as the current event
        procedureProgress:
          type: number
          description: Percentage progress of the procedure
    ReportEventReadDto:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
        address:
          type: string
        description:
          type: string
        type:
          $ref: '#/components/schemas/Type'
        criticality:
          $ref: '#/components/schemas/Criticality'
        equipment:
          $ref: '#/components/schemas/EquipmentReadDto'
        status:
          $ref: '#/components/schemas/Status'
        category:
          $ref: '#/components/schemas/ReportCategory'
        lastModificationDate:
          $ref: '#/components/schemas/DateTime'
        creationDate:
          $ref: '#/components/schemas/DateTime'
        closeDate:
          $ref: '#/components/schemas/DateTime'
        externalSourceRef:
          type: string
        procedureId:
          $ref: '#/components/schemas/UUID'
        linkedEvents:
          type: number
          description: Number of events that have the same procedure as the current event
        procedureProgress:
          type: number
          description: Percentage progress of the procedure
    OperatorEventReadDto:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
        address:
          type: string
        description:
          type: string
        type:
          $ref: '#/components/schemas/Type'
        criticality:
          $ref: '#/components/schemas/Criticality'
        equipment:
          $ref: '#/components/schemas/EquipmentReadDto'
        status:
          $ref: '#/components/schemas/Status'
        category:
          $ref: '#/components/schemas/OperatorCategory'
        lastModificationDate:
          $ref: '#/components/schemas/DateTime'
        creationDate:
          $ref: '#/components/schemas/DateTime'
        closeDate:
          $ref: '#/components/schemas/DateTime'
        endDate:
          $ref: '#/components/schemas/DateTime'
        startDate:
          $ref: '#/components/schemas/DateTime'
        externalSourceRef:
          type: string
        procedureId:
          $ref: '#/components/schemas/UUID'
        linkedEvents:
          type: number
          description: Number of events that have the same procedure as the current event
        procedureProgress:
          type: number
          description: Percentage progress of the procedure
    ProcedureReadDto:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
        lastModificationDate:
          $ref: '#/components/schemas/DateTime'
        creationDate:
          $ref: '#/components/schemas/DateTime'
        progress:
          type: number
        events:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/OperatorEventReadDto'
              - $ref: '#/components/schemas/AlertEventReadDto'
              - $ref: '#/components/schemas/ReportEventReadDto'
        actions:
          type: array
          items:
            $ref: '#/components/schemas/ActionReadDto'
    ProcedureWriteDto:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
        events:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/OperatorEventWriteDto'
              - $ref: '#/components/schemas/AlertEventWriteDto'
              - $ref: '#/components/schemas/ReportEventWriteDto'
        actions:
          type: array
          items:
            $ref: '#/components/schemas/ActionReadDto'
...